refactor(exec): move WeightCaches + FP8CacheEntry to exec/weight_caches.h
